fix: Use more informative patch error if the same APK is patched twice

+++ patches/src/main/kotlin/app/morphe/patches/shared/layout/branding/AddBrandLicensePatch.kt
@@ -1,5 +1,6 @@
 package app.morphe.patches.shared.layout.branding
 
+import app.morphe.patcher.patch.PatchException
 import app.morphe.patcher.patch.rawResourcePatch
 import app.morphe.util.inputStreamFromBundledResource
 import java.nio.file.Files
@@ -17,6 +18,12 @@ internal val addBrandLicensePatch = rawResourcePatch {
         )!!
 
         val targetFile = get(sourceFileName, false).toPath()
+
+        // Check if target file exists and give a more informative error
+        // because Files.copy throws an exception if the file already exists.
+        if (Files.exists(targetFile)) {
+            throw PatchException("App is already modified with Morphe patches ($targetFile already exists)")
+        }
 
         Files.copy(inputFileStream, targetFile)
     }
